Installateur FS V1.0
Refonte totale du précèdent installateur de FS.

- Amélioration majeure des performances.

- Amélioration de l'ergonomie, deux clics suffisent dans la majorité des
cas pour configurer l'application.

- Amélioration de l'apparence.

+++ public/index.php
@@ -37,6 +37,10 @@ try {
     echo $application->handle()->getContent();
 
 } catch (\Exception $e) {
-    echo $e->getMessage() . '<br>';
-    echo '<pre>' . $e->getTraceAsString() . '</pre>';
+    if(file_exists("installer.php")) {
+        include "installer.php";
+    }else {
+        echo $e->getMessage() . '<br>';
+        echo '<pre>' . $e->getTraceAsString() . '</pre>';
+    }
 }
